+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Path IO 0.2.0
+
+* Added functions from `temporary`: `withTempFile`, `withTempDir`,
+  `withSystemTempFile`, `withSystemTempDir`, `openTempFile`,
+  `openBinaryTempFile`, and `createTempDir`. `temporary` is a lightweight
+  and ubiquitous package, so depending on it should be OK.
+
 ## Path IO 0.1.1
 
 * Fixed type signatures of `renameFile` and `copyFile`.